requirements.txt
openai >= 1.51.2
streamlit >= 1.21.0
python-dotenv >= 1.0.0